取消service接口
