+++ weather-city/src/main/java/com/istimeless/weathercity/service/CacheCityService.java
@@ -1,9 +1,32 @@
 package com.istimeless.weathercity.service;
 
+import com.istimeless.weathercity.advice.WeatherCityEnum;
+import com.istimeless.weathercity.vo.CityRequest;
+import com.istimeless.weathercity.vo.CityResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
 /**
  * @author lijiayin
  */
-public interface CacheCityService {
+@Slf4j
+@Service
+public class CacheCityService {
     
-    void cacheCity();
+    private final CityInfoService cityInfoService;
+    
+    private final SaveCityInfoService saveCityInfoService;
+
+    public CacheCityService(CityInfoService cityInfoService,
+                            SaveCityInfoService saveCityInfoService) {
+        this.cityInfoService = cityInfoService;
+        this.saveCityInfoService = saveCityInfoService;
+    }
+    
+    public void cacheCity() {
+        CityResponse cityResponse = cityInfoService.cityInfo(CityRequest.builder().subdistrict(3).build());
+        if(WeatherCityEnum.SUCCESS.getCode().equals(cityResponse.getStatus())){
+            saveCityInfoService.saveCityInfo(cityResponse);
+        }
+    }
 }
