+++ weather-city/src/test/java/com/istimeless/weathercity/service/CacheCityServiceTest.java
@@ -1,13 +1,11 @@
-package com.istimeless.weathercity.service.impl;
+package com.istimeless.weathercity.service;
 
 import com.istimeless.weathercity.WeatherCityApplicationTests;
 import com.istimeless.weathercity.service.CacheCityService;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import static org.junit.Assert.*;
-
-public class CacheCityServiceImplTest extends WeatherCityApplicationTests {
+public class CacheCityServiceTest extends WeatherCityApplicationTests {
 
     @Autowired
     private CacheCityService cacheCityService;
